refactor : 각 클래스 패키지 구조 변경 및 클래스명 변경 MessageType을 MessageProvider로 변경
