+++ src/main/java/calculator/domain/message/MessageProvider.java
@@ -1,0 +1,5 @@
+package calculator.domain.message;
+
+public interface MessageProvider {
+    String getMessage();
+}
