--- src/main/java/calculator/message/MessageType.java
@@ -1,5 +1,0 @@
-package calculator.message;
-
-public interface MessageType {
-    String getMessage();
-}
